backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
flask-restful==0.3.10
mongoengine==0.28.0
flask-jwt-extended==4.6.0
flask-mail==0.9.1
flask-cors==4.0.0
gunicorn==21.2.0
pymongo[srv]==4.6.1
dnspython==2.4.2
Werkzeug==3.0.1
